requirements.txt
streamlit==1.48.0
langchain== 0.3.27
langchain-core== 0.3.74
langchain-google-genai==2.0.10
langchain-community==0.3.27
google-generativeai==0.8.5
python-dotenv==1.0.0
pypdf==3.17.1
faiss-cpu==1.11.0.post1